api-service/requirements.txt
# my_project/api-service/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
mysql-connector-python==8.0.33

